+++ src/Changes/SelectionSet/RenameSelectionSetChange.ts
@@ -1,6 +1,6 @@
 import {Change} from '@zeainc/zea-ux'
 import {EventEmitter} from "@zeainc/zea-engine";
-import {SelectionSet} from "../SelectionSet";
+import {SelectionSet} from "../../SelectionSet";
 
 export default class RenameSelectionSetChange extends Change {
   selectionSet: SelectionSet
