+++ src/Changes/SelectionSetChanges.ts
@@ -1,0 +1,4 @@
+export { CreateSelectionSetChange } from './SelectionSet/CreateSelectionSetChange'
+export { DeleteSelectionSetChange } from './SelectionSet/DeleteSelectionSetChange'
+export { RenameSelectionSetChange } from './SelectionSet/RenameSelectionSetChange'
+export { UpdateSelectionSetChange } from './SelectionSet/UpdateSelectionSetChange'
