+++ src/Changes/ViewChanges.ts
@@ -1,0 +1,4 @@
+export { CreateViewChange } from './View/CreateViewChange'
+export { DeleteViewChange } from './View/DeleteViewChange'
+export { RenameViewChange } from './View/RenameViewChange'
+export { UpdateViewCameraChange } from './View/UpdateViewCameraChange'
